+++ com.riccardomalavolti.apps.todolist/src/main/java/com/riccardomalavolti/apps/todolist/Main.java
@@ -1,15 +1,14 @@
-package com.riccardomalavolti.apps.todolist.main;
+package com.riccardomalavolti.apps.todolist;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-
 public class Main {
 
 	private static final Logger LOGGER = LogManager.getLogger(Main.class);
-	
+
 	public static void main(String[] args) {
-		
+
 		LOGGER.info("The application is starting.");
 	}
